Create shared helper to calculate competition standings
