+++ WebService/CreateDatabaseTables.sql
@@ -5,7 +5,7 @@
 -- Dumped from database version 16.2 (Debian 16.2-1.pgdg120+2)
 -- Dumped by pg_dump version 16.0
 
--- Started on 2024-04-18 07:45:16 EDT
+-- Started on 2024-04-27 20:50:16 EDT
 
 SET statement_timeout = 0;
 SET lock_timeout = 0;
@@ -53,7 +53,8 @@ CREATE TABLE public.competitions (
     access_token text NOT NULL,
     iana_timezone text NOT NULL,
     competition_id uuid NOT NULL,
-    admin_user_id bytea NOT NULL
+    admin_user_id bytea NOT NULL,
+    state smallint DEFAULT 0 NOT NULL
 );
 
 
@@ -152,7 +153,7 @@ CREATE TABLE public.workouts (
 ALTER TABLE public.workouts OWNER TO postgres;
 
 --
--- TOC entry 3244 (class 2606 OID 16428)
+-- TOC entry 3245 (class 2606 OID 16428)
 -- Name: push_tokens Primary Key; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -161,7 +162,7 @@ ALTER TABLE ONLY public.push_tokens
 
 
 --
--- TOC entry 3238 (class 2606 OID 16430)
+-- TOC entry 3239 (class 2606 OID 16430)
 -- Name: oauth_clients oauth_clients_pkey; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -170,7 +171,7 @@ ALTER TABLE ONLY public.oauth_clients
 
 
 --
--- TOC entry 3236 (class 2606 OID 16432)
+-- TOC entry 3237 (class 2606 OID 16432)
 -- Name: competitions primary_key; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -179,7 +180,7 @@ ALTER TABLE ONLY public.competitions
 
 
 --
--- TOC entry 3233 (class 2606 OID 16434)
+-- TOC entry 3234 (class 2606 OID 16434)
 -- Name: activity_summaries primary_key2; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -188,7 +189,7 @@ ALTER TABLE ONLY public.activity_summaries
 
 
 --
--- TOC entry 3242 (class 2606 OID 16436)
+-- TOC entry 3243 (class 2606 OID 16436)
 -- Name: oauth_tokens token_primary_key; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -197,7 +198,7 @@ ALTER TABLE ONLY public.oauth_tokens
 
 
 --
--- TOC entry 3250 (class 2606 OID 16438)
+-- TOC entry 3251 (class 2606 OID 16438)
 -- Name: users_competitions users_competitions_pkey; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -206,7 +207,7 @@ ALTER TABLE ONLY public.users_competitions
 
 
 --
--- TOC entry 3246 (class 2606 OID 16440)
+-- TOC entry 3247 (class 2606 OID 16440)
 -- Name: users users_pkey; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -215,7 +216,7 @@ ALTER TABLE ONLY public.users
 
 
 --
--- TOC entry 3252 (class 2606 OID 16442)
+-- TOC entry 3253 (class 2606 OID 16442)
 -- Name: workouts workouts_pkey; Type: CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -224,7 +225,7 @@ ALTER TABLE ONLY public.workouts
 
 
 --
--- TOC entry 3234 (class 1259 OID 16443)
+-- TOC entry 3235 (class 1259 OID 16443)
 -- Name: fki_admin_user_id_foreignkey; Type: INDEX; Schema: public; Owner: postgres
 --
 
@@ -232,7 +233,7 @@ CREATE INDEX fki_admin_user_id_foreignkey ON public.competitions USING btree (ad
 
 
 --
--- TOC entry 3247 (class 1259 OID 16444)
+-- TOC entry 3248 (class 1259 OID 16444)
 -- Name: fki_competition_id_foreignkey; Type: INDEX; Schema: public; Owner: postgres
 --
 
@@ -240,7 +241,7 @@ CREATE INDEX fki_competition_id_foreignkey ON public.users_competitions USING bt
 
 
 --
--- TOC entry 3239 (class 1259 OID 16445)
+-- TOC entry 3240 (class 1259 OID 16445)
 -- Name: fki_oauth_tokens_client_id_fkey; Type: INDEX; Schema: public; Owner: postgres
 --
 
@@ -248,7 +249,7 @@ CREATE INDEX fki_oauth_tokens_client_id_fkey ON public.oauth_tokens USING btree 
 
 
 --
--- TOC entry 3240 (class 1259 OID 16446)
+-- TOC entry 3241 (class 1259 OID 16446)
 -- Name: fki_oauth_tokens_user_id_fkey; Type: INDEX; Schema: public; Owner: postgres
 --
 
@@ -256,7 +257,7 @@ CREATE INDEX fki_oauth_tokens_user_id_fkey ON public.oauth_tokens USING btree (u
 
 
 --
--- TOC entry 3248 (class 1259 OID 16447)
+-- TOC entry 3249 (class 1259 OID 16447)
 -- Name: fki_user_id_foreignkey; Type: INDEX; Schema: public; Owner: postgres
 --
 
@@ -264,7 +265,7 @@ CREATE INDEX fki_user_id_foreignkey ON public.users_competitions USING btree (us
 
 
 --
--- TOC entry 3254 (class 2606 OID 16448)
+-- TOC entry 3255 (class 2606 OID 16448)
 -- Name: competitions admin_user_id_foreignkey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -273,7 +274,7 @@ ALTER TABLE ONLY public.competitions
 
 
 --
--- TOC entry 3258 (class 2606 OID 16453)
+-- TOC entry 3259 (class 2606 OID 16453)
 -- Name: users_competitions competition_id_foreignkey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -282,7 +283,7 @@ ALTER TABLE ONLY public.users_competitions
 
 
 --
--- TOC entry 3255 (class 2606 OID 16458)
+-- TOC entry 3256 (class 2606 OID 16458)
 -- Name: oauth_tokens oauth_tokens_client_id_fkey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -291,7 +292,7 @@ ALTER TABLE ONLY public.oauth_tokens
 
 
 --
--- TOC entry 3256 (class 2606 OID 16463)
+-- TOC entry 3257 (class 2606 OID 16463)
 -- Name: oauth_tokens oauth_tokens_user_id_fkey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -300,7 +301,7 @@ ALTER TABLE ONLY public.oauth_tokens
 
 
 --
--- TOC entry 3257 (class 2606 OID 16468)
+-- TOC entry 3258 (class 2606 OID 16468)
 -- Name: push_tokens userIdFKey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -309,7 +310,7 @@ ALTER TABLE ONLY public.push_tokens
 
 
 --
--- TOC entry 3260 (class 2606 OID 16473)
+-- TOC entry 3261 (class 2606 OID 16473)
 -- Name: workouts userId_fkey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -318,7 +319,7 @@ ALTER TABLE ONLY public.workouts
 
 
 --
--- TOC entry 3259 (class 2606 OID 16478)
+-- TOC entry 3260 (class 2606 OID 16478)
 -- Name: users_competitions user_id_foreignkey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -327,7 +328,7 @@ ALTER TABLE ONLY public.users_competitions
 
 
 --
--- TOC entry 3253 (class 2606 OID 16483)
+-- TOC entry 3254 (class 2606 OID 16483)
 -- Name: activity_summaries user_id_foreignkey; Type: FK CONSTRAINT; Schema: public; Owner: postgres
 --
 
@@ -335,7 +336,7 @@ ALTER TABLE ONLY public.activity_summaries
     ADD CONSTRAINT user_id_foreignkey FOREIGN KEY (user_id) REFERENCES public.users(user_id) ON DELETE CASCADE;
 
 
--- Completed on 2024-04-18 07:45:17 EDT
+-- Completed on 2024-04-27 20:50:16 EDT
 
 --
 -- PostgreSQL database dump complete
