+++ WebService/FitWithFriends/sql/competitions.sql
@@ -11,7 +11,7 @@ ON CONFLICT (user_id, competition_id) DO NOTHING;
 
 /* @name GetCompetition */
 /* Does not return the access_token field - we will only return that to admin users */
-SELECT start_date, end_date, display_name, admin_user_id, iana_timezone, competition_id FROM competitions WHERE competition_id = :competitionId!;
+SELECT start_date, end_date, display_name, admin_user_id, iana_timezone, competition_id, state FROM competitions WHERE competition_id = :competitionId!;
 
 /* @name GetCompetitionAdminDetails */
 SELECT start_date, end_date, display_name, admin_user_id, access_token, iana_timezone, competition_id FROM competitions WHERE competition_id = :competitionId! AND admin_user_id = :adminUserId!;
