+++ WebService/FitWithFriends/utilities/enums/FWFErrorCodes.ts
