+++ WebService/FitWithFriends/utilities/enums/PushNotificationPlatform.ts
